Added onClick event for table click, trying to get redux to rerender full page
