+++ src/components/instructor/InstructorFig1Table.js
@@ -1,356 +1,214 @@
-
 import React from 'react';
-import classNames from 'classnames';
 import PropTypes from 'prop-types';
 import { withStyles } from '@material-ui/core/styles';
 import Table from '@material-ui/core/Table';
 import TableBody from '@material-ui/core/TableBody';
 import TableCell from '@material-ui/core/TableCell';
 import TableHead from '@material-ui/core/TableHead';
-import TablePagination from '@material-ui/core/TablePagination';
 import TableRow from '@material-ui/core/TableRow';
-import TableSortLabel from '@material-ui/core/TableSortLabel';
-import Toolbar from '@material-ui/core/Toolbar';
-import Typography from '@material-ui/core/Typography';
 import Paper from '@material-ui/core/Paper';
-import Checkbox from '@material-ui/core/Checkbox';
-import IconButton from '@material-ui/core/IconButton';
-import Tooltip from '@material-ui/core/Tooltip';
-import DeleteIcon from '@material-ui/icons/Delete';
-import FilterListIcon from '@material-ui/icons/FilterList';
-import { lighten } from '@material-ui/core/styles/colorManipulator';
+import { api_endpoint } from '../../constants.js';
+import obj from '../MobileTools.js'
+import lists from '../../course_instructor_list.json';
+import { connect } from 'react-redux';
+import {
+  setSearchStatus,
+  SearchStatus,
+  setSearchType,
+  setSearchText,
+  SearchType,
+} from '../../actions';
 
-let counter = 0;
-function createData(name, calories, fat, carbs, protein) {
-  counter += 1;
-  return { id: counter, name, calories, fat, carbs, protein };
-}
+// Get course and instructor lists
+const course_list = lists['courses']
+// const instructor_list = lists['instructors']
 
-function desc(a, b, orderBy) {
-  if (b[orderBy] < a[orderBy]) {
-    return -1;
+// Define mobile parameters
+var em = obj['em']
+var mobile = obj['mobile']
+var head_text_size = (em/16).toString();
+var table_padding = 3;
+if (mobile) {
+  head_text_size = (em/6).toString()
+  table_padding = 1.25}
+
+const CustomTableCell = withStyles(theme => ({
+  head: {
+    backgroundColor: '#841617',
+    color: theme.palette.common.white,
+    fontSize: head_text_size+'rem',
+    fontWeight: 'bold',
+    padding: table_padding*theme.spacing.unit,
+    paddingTop: 0.5*table_padding*theme.spacing.unit,
+    paddingBottom: 0.5*table_padding*theme.spacing.unit
+
+  },
+  body: {
+    padding: table_padding*theme.spacing.unit,
+    paddingTop: 0.5*table_padding*theme.spacing.unit,
+    paddingBottom: 0.5*table_padding*theme.spacing.unit,
+}}))(TableCell);
+
+const CustomTableCellHyperlink = withStyles(theme => ({
+  body: {
+    '&:hover': {
+    fontWeight:'bold',
+    textDecoration: 'underline',
+    cursor:'pointer'
   }
-  if (b[orderBy] > a[orderBy]) {
-    return 1;
-  }
-  return 0;
-}
+}}))(CustomTableCell);
 
-function stableSort(array, cmp) {
-  const stabilizedThis = array.map((el, index) => [el, index]);
-  stabilizedThis.sort((a, b) => {
-    const order = cmp(a[0], b[0]);
-    if (order !== 0) return order;
-    return a[1] - b[1];
-  });
-  return stabilizedThis.map(el => el[0]);
-}
-
-function getSorting(order, orderBy) {
-  return order === 'desc' ? (a, b) => desc(a, b, orderBy) : (a, b) => -desc(a, b, orderBy);
-}
-
-const rows = [
-  { id: 'name', numeric: false, disablePadding: true, label: 'Dessert (100g serving)' },
-  { id: 'calories', numeric: true, disablePadding: false, label: 'Calories' },
-  { id: 'fat', numeric: true, disablePadding: false, label: 'Fat (g)' },
-  { id: 'carbs', numeric: true, disablePadding: false, label: 'Carbs (g)' },
-  { id: 'protein', numeric: true, disablePadding: false, label: 'Protein (g)' },
-];
-
-class EnhancedTableHead extends React.Component {
-  createSortHandler = property => event => {
-    this.props.onRequestSort(event, property);
-  };
-
-  render() {
-    const { onSelectAllClick, order, orderBy, numSelected, rowCount } = this.props;
-
-    return (
-      <TableHead>
-        <TableRow>
-          <TableCell padding="checkbox">
-            <Checkbox
-              indeterminate={numSelected > 0 && numSelected < rowCount}
-              checked={numSelected === rowCount}
-              onChange={onSelectAllClick}
-            />
-          </TableCell>
-          {rows.map(
-            row => (
-              <TableCell
-                key={row.id}
-                align={row.numeric ? 'right' : 'left'}
-                padding={row.disablePadding ? 'none' : 'default'}
-                sortDirection={orderBy === row.id ? order : false}
-              >
-                <Tooltip
-                  title="Sort"
-                  placement={row.numeric ? 'bottom-end' : 'bottom-start'}
-                  enterDelay={300}
-                >
-                  <TableSortLabel
-                    active={orderBy === row.id}
-                    direction={order}
-                    onClick={this.createSortHandler(row.id)}
-                  >
-                    {row.label}
-                  </TableSortLabel>
-                </Tooltip>
-              </TableCell>
-            ),
-            this,
-          )}
-        </TableRow>
-      </TableHead>
-    );
-  }
-}
-
-EnhancedTableHead.propTypes = {
-  numSelected: PropTypes.number.isRequired,
-  onRequestSort: PropTypes.func.isRequired,
-  onSelectAllClick: PropTypes.func.isRequired,
-  order: PropTypes.string.isRequired,
-  orderBy: PropTypes.string.isRequired,
-  rowCount: PropTypes.number.isRequired,
-};
-
-const toolbarStyles = theme => ({
-  root: {
-    paddingRight: theme.spacing.unit,
-  },
-  highlight:
-    theme.palette.type === 'light'
-      ? {
-          color: theme.palette.secondary.main,
-          backgroundColor: lighten(theme.palette.secondary.light, 0.85),
-        }
-      : {
-          color: theme.palette.text.primary,
-          backgroundColor: theme.palette.secondary.dark,
-        },
-  spacer: {
-    flex: '1 1 100%',
-  },
-  actions: {
-    color: theme.palette.text.secondary,
-  },
-  title: {
-    flex: '0 0 auto',
-  },
-});
-
-let EnhancedTableToolbar = props => {
-  const { numSelected, classes } = props;
-
-  return (
-    <Toolbar
-      className={classNames(classes.root, {
-        [classes.highlight]: numSelected > 0,
-      })}
-    >
-      <div className={classes.title}>
-        {numSelected > 0 ? (
-          <Typography color="inherit" variant="subtitle1">
-            {numSelected} selected
-          </Typography>
-        ) : (
-          <Typography variant="h6" id="tableTitle">
-            Nutrition
-          </Typography>
-        )}
-      </div>
-      <div className={classes.spacer} />
-      <div className={classes.actions}>
-        {numSelected > 0 ? (
-          <Tooltip title="Delete">
-            <IconButton aria-label="Delete">
-              <DeleteIcon />
-            </IconButton>
-          </Tooltip>
-        ) : (
-          <Tooltip title="Filter list">
-            <IconButton aria-label="Filter list">
-              <FilterListIcon />
-            </IconButton>
-          </Tooltip>
-        )}
-      </div>
-    </Toolbar>
-  );
-};
-
-EnhancedTableToolbar.propTypes = {
-  classes: PropTypes.object.isRequired,
-  numSelected: PropTypes.number.isRequired,
-};
-
-EnhancedTableToolbar = withStyles(toolbarStyles)(EnhancedTableToolbar);
-
+// This defines styles for the table
 const styles = theme => ({
   root: {
+    align: 'center',
     width: '100%',
-    marginTop: theme.spacing.unit * 3,
+    overflowX: 'auto',
   },
   table: {
-    minWidth: 1020,
   },
-  tableWrapper: {
-    overflowX: 'auto',
+  tableRow: {
+    padding: theme.spacing.unit,
+    '&:nth-of-type(odd)': {
+      backgroundColor: theme.palette.background.default, // Might want to change this if desired
+    },
+    '&:hover': {
+      backgroundColor: '#f3b7b7',
+    },
   },
 });
 
-class EnhancedTable extends React.Component {
-  state = {
-    order: 'asc',
-    orderBy: 'calories',
-    selected: [],
-    data: [
-      createData('Cupcake', 305, 3.7, 67, 4.3),
-      createData('Donut', 452, 25.0, 51, 4.9),
-      createData('Eclair', 262, 16.0, 24, 6.0),
-      createData('Frozen yoghurt', 159, 6.0, 24, 4.0),
-      createData('Gingerbread', 356, 16.0, 49, 3.9),
-      createData('Honeycomb', 408, 3.2, 87, 6.5),
-      createData('Ice cream sandwich', 237, 9.0, 37, 4.3),
-      createData('Jelly Bean', 375, 0.0, 94, 0.0),
-      createData('KitKat', 518, 26.0, 65, 7.0),
-      createData('Lollipop', 392, 0.2, 98, 0.0),
-      createData('Marshmallow', 318, 0, 81, 2.0),
-      createData('Nougat', 360, 19.0, 9, 37.0),
-      createData('Oreo', 437, 18.0, 63, 4.0),
-    ],
-    page: 0,
-    rowsPerPage: 5,
-  };
+// This is the function that will fetch the desired data from the api
+const API = api_endpoint + 'instructors/';
 
-  handleRequestSort = (event, property) => {
-    const orderBy = property;
-    let order = 'desc';
+class InstructorFig1Table extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = { loadedAPI: false, data: [], uuid: props.uuid };
+  }
 
-    if (this.state.orderBy === property && this.state.order === 'desc') {
-      order = 'asc';
-    }
-
-    this.setState({ order, orderBy });
-  };
-
-  handleSelectAllClick = event => {
-    if (event.target.checked) {
-      this.setState(state => ({ selected: state.data.map(n => n.id) }));
-      return;
-    }
-    this.setState({ selected: [] });
-  };
-
-  handleClick = (event, id) => {
-    const { selected } = this.state;
-    const selectedIndex = selected.indexOf(id);
-    let newSelected = [];
-
-    if (selectedIndex === -1) {
-      newSelected = newSelected.concat(selected, id);
-    } else if (selectedIndex === 0) {
-      newSelected = newSelected.concat(selected.slice(1));
-    } else if (selectedIndex === selected.length - 1) {
-      newSelected = newSelected.concat(selected.slice(0, -1));
-    } else if (selectedIndex > 0) {
-      newSelected = newSelected.concat(
-        selected.slice(0, selectedIndex),
-        selected.slice(selectedIndex + 1),
-      );
-    }
-
-    this.setState({ selected: newSelected });
-  };
-
-  handleChangePage = (event, page) => {
-    this.setState({ page });
-  };
-
-  handleChangeRowsPerPage = event => {
-    this.setState({ rowsPerPage: event.target.value });
-  };
-
-  isSelected = id => this.state.selected.indexOf(id) !== -1;
+  componentDidMount() {
+    // This will call the api when the component "Mounts", i.e. when the page is accessed
+    fetch(API + this.state.uuid.toString() + '/figure1')
+      .then(response => response.json())
+      .then(data => this.setState({ data: data.result, loadedAPI: true }));
+  }
 
   render() {
-    const { classes } = this.props;
-    const { data, order, orderBy, selected, rowsPerPage, page } = this.state;
-    const emptyRows = rowsPerPage - Math.min(rowsPerPage, data.length - page * rowsPerPage);
+    if (!this.state.loadedAPI) {
+      return null;
+    } else {
 
-    return (
-      <Paper className={classes.root}>
-        <EnhancedTableToolbar numSelected={selected.length} />
-        <div className={classes.tableWrapper}>
-          <Table className={classes.table} aria-labelledby="tableTitle">
-            <EnhancedTableHead
-              numSelected={selected.length}
-              order={order}
-              orderBy={orderBy}
-              onSelectAllClick={this.handleSelectAllClick}
-              onRequestSort={this.handleRequestSort}
-              rowCount={data.length}
-            />
-            <TableBody>
-              {stableSort(data, getSorting(order, orderBy))
-                .slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
-                .map(n => {
-                  const isSelected = this.isSelected(n.id);
-                  return (
-                    <TableRow
-                      hover
-                      onClick={event => this.handleClick(event, n.id)}
-                      role="checkbox"
-                      aria-checked={isSelected}
-                      tabIndex={-1}
-                      key={n.id}
-                      selected={isSelected}
-                    >
-                      <TableCell padding="checkbox">
-                        <Checkbox checked={isSelected} />
-                      </TableCell>
-                      <TableCell component="th" scope="row" padding="none">
-                        {n.name}
-                      </TableCell>
-                      <TableCell align="right">{n.calories}</TableCell>
-                      <TableCell align="right">{n.fat}</TableCell>
-                      <TableCell align="right">{n.carbs}</TableCell>
-                      <TableCell align="right">{n.protein}</TableCell>
-                    </TableRow>
-                  );
-                })}
-              {emptyRows > 0 && (
-                <TableRow style={{ height: 49 * emptyRows }}>
-                  <TableCell colSpan={6} />
-                </TableRow>
-              )}
-            </TableBody>
-          </Table>
-        </div>
-        <TablePagination
-          rowsPerPageOptions={[5, 10, 25]}
-          component="div"
-          count={data.length}
-          rowsPerPage={rowsPerPage}
-          page={page}
-          backIconButtonProps={{
-            'aria-label': 'Previous Page',
-          }}
-          nextIconButtonProps={{
-            'aria-label': 'Next Page',
-          }}
-          onChangePage={this.handleChangePage}
-          onChangeRowsPerPage={this.handleChangeRowsPerPage}
-        />
-      </Paper>
-    );
+      let MyTable = withStyles(styles)(CustomizedTable); // This is important
+      // Get the data ready to pass to the table by rounding and adding ids
+      var table_data = this.state.data;
+      table_data.courses.forEach((item, i) => {
+        item['instr_rating_in_course'] = item['instr_rating_in_course'].toFixed(
+          2
+        ); // Convert the long floats to 2 decimal places
+        item['display name'] =
+          item['dept name'] +
+          item['course number'].toString() +
+          ': ' +
+          item['course name'];
+        item['id'] = i + 1;
+      });
+      return <MyTable data={table_data} />;
+    }
   }
 }
 
-EnhancedTable.propTypes = {
+//This is the function to create the table for figure 1
+class CustomizedTable extends React.Component {
+  constructor(props){
+    super(props)
+  this.state = {classes:props.classes, data:props.data, rows:props.data.courses}
+  }
+
+handleClick(event, id){
+  var clicked = this.state.rows[id-1]['display name']
+  // Convert course list to dict/hash
+  var course_dict = course_list.reduce((obj, item) => {
+     obj[item['label'].toString()] = item['value'].toString()
+     return obj
+   }, {})
+  // Get the course uuid from the course dict
+  var course_uuid = course_dict[clicked]
+  if (typeof course_uuid != 'undefined') {
+  // Now just need to pass to DataView , which will trigger redux state change
+  setSearchType('COURSE');
+  setSearchText(course_uuid);
+  this.setState(this.state)
+}}
+
+render() {
+  var data = this.state.data
+  var classes = this.state.classes
+  var rows = this.state.rows
+  return (
+    <div>
+      <h1 className='title' style={{paddingBottom: '0.5em'}}>
+        {' '}
+        {data['instructor name']}
+      </h1>
+      <h2 className='subtitle'>
+        {' '}
+        {data['instructor name']} has taught these courses in the <b>previous 3 years</b>
+      </h2>
+      <Paper className={classes.root}>
+        <Table className={classes.table}>
+          <TableHead>
+            <TableRow className={classes.tableRow}>
+              <CustomTableCell align='left'>
+
+                Course Name
+              </CustomTableCell>
+              <CustomTableCell
+                align="center">
+                Average Course Rating (1-5)
+              </CustomTableCell>
+              <CustomTableCell
+                align="center">
+                {data['instructor name']} Rating in Course (1-5)
+              </CustomTableCell>
+              <CustomTableCell
+                align="right"
+              >
+                Semester(s) Taught
+              </CustomTableCell>
+            </TableRow>
+          </TableHead>
+          <TableBody>
+            {rows.map(row => (
+              <TableRow className={classes.tableRow} key={row.id}>
+                <CustomTableCellHyperlink component="th" scope="row" onClick={event => this.handleClick(event, row.id)}>
+                  {row['display name']}
+                </CustomTableCellHyperlink>
+                <CustomTableCell align='center'>
+                {row['avg_course_rating'].toFixed(2).toString()}
+                </CustomTableCell>
+                <CustomTableCell align="center">
+                  {row['instr_rating_in_course']}
+                </CustomTableCell>
+                <CustomTableCell align="right">{row['term']}</CustomTableCell>
+              </TableRow>
+            ))}
+          </TableBody>
+        </Table>
+      </Paper>
+    </div>
+  );
+}}
+
+CustomizedTable.propTypes = {
   classes: PropTypes.object.isRequired,
 };
 
-export default withStyles(styles)(EnhancedTable);
+const mapStateToProps = state => {
+  return {
+    search_type: state.search_type,
+    search_text: state.search_text
+  };
+};
+
+export default (InstructorFig1Table);
